requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pythontextnow-1.1.0-py3-none-any (1).whl
